requirements.txt
streamlit
google-generativeai
PyMuPDF
Pillow
pandas
msal
requests
python-dotenv
pdf2image
